Changed formatting and conducted tests
Changed the formatting for the mobile page, Fixing radio button issue
and adding some additional styling.

Added a logout button. This nullifies the session variable "Username"
to effectively limit the user's access to all pages but the login page.

Added code to desktop pages to check if a request is coming from a
mobile browser. If so, the user will be redirected to the mobile site.
Also added code to do the opposite to the mobile pages, but it is
commented out for testing purposes