Altered signature functionality, added to desktop
Mobile signature is now creating an image and storing it
in /signatures. The image name is being created as the current
datetime converted to an integer and with 2 random digits assigned.

This was a necessary change because the json was large to store,
and difficult to regenerate. In addition, it would have likely
been difficult to add the image to the pdf file.

The database is now storing a path to the image relative to the
virtual directory of the application. This will allow the
desktop site to render the image seamlessly. Tested all
functionality, and it appears to be working.

My only concern is whether this functionality will work on our
server since we may not have rights to create a file. If so, not
sure how to address it.